Move client/plrdlg.* to client/views with a rename
